refactor: modularized menu files
