+++ app/store/index.js
@@ -1,6 +1,5 @@
 import { combineReducers, createStore, applyMiddleware } from 'redux';
 import thunkMiddleware from 'redux-thunk';
-import { createLogger } from 'redux-logger';
 import user from './user';
 
 const rootReducer = combineReducers({
@@ -8,7 +7,7 @@ const rootReducer = combineReducers({
 });
 const store = createStore(
   rootReducer,
-  applyMiddleware(thunkMiddleware, createLogger())
+  applyMiddleware(thunkMiddleware)
 );
 
 export default store;
